requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
watchdog>=2.2.0
scikit-learn>=1.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pillow>=9.0.0
numpy>=1.24.0
joblib>=1.2.0
